change: 3.3.6で使用したスクリプト（一部修正）

+++ shells/createNC3Package.sh
@@ -20,34 +20,38 @@ echo "+----------------------------------------+"
 
 execute "cd $PKGDIR"
 
-if [ -d $PROJECTNAME ]; then
-	execute "rm -Rf $PROJECTNAME"
-fi
-if [ -d $GITZIPFILE ]; then
-	execute "rm -Rf $GITZIPFILE"
-fi
-if [ -f $NC3VERSION.zip ]; then
+if [ "$MODE" = "prod" ]; then
+	if [ -d $PROJECTNAME ]; then
+		execute "rm -Rf $PROJECTNAME"
+	fi
+	if [ -d $GITZIPFILE ]; then
+		execute "rm -Rf $GITZIPFILE"
+	fi
+	if [ -f $NC3VERSION.zip ]; then
+		execute "rm -f $NC3VERSION.zip"
+	fi
+
+	execute "wget $GITURL/$PROJECTNAME/archive/$NC3VERSION.zip"
+
+	if [ ! -f $NC3VERSION.zip ]; then
+		echo "$NC3VERSION.zipファイルが存在しません"
+		exit 1
+	fi
+
+	execute "unzip $NC3VERSION.zip"
 	execute "rm -f $NC3VERSION.zip"
+	execute "mv $GITZIPFILE $PROJECTNAME"
+
+	execute "cd $PROJECTNAME"
+	execute "${CMDCMPOSER} install --no-dev"
+	execute "rm -f app/tmp/logs/debug.log"
+	execute "rm -f app/tmp/cache/persistent/myapp_*"
+
+	execute "cd $PKGDIR"
+	execute "zip -r $ZIPFILE.zip $PROJECTNAME"
+else
+	execute "cp -Rpf $WORKDIR/$PROJECTNAME $PKGDIR/$PROJECTNAME"
 fi
-
-execute "wget $GITURL/$PROJECTNAME/archive/$NC3VERSION.zip"
-
-if [ ! -f $NC3VERSION.zip ]; then
-	echo "$NC3VERSION.zipファイルが存在しません"
-	exit 1
-fi
-
-execute "unzip $NC3VERSION.zip"
-execute "rm -f $NC3VERSION.zip"
-execute "mv $GITZIPFILE $PROJECTNAME"
-
-execute "cd $PROJECTNAME"
-execute "${CMDCMPOSER} install --no-dev"
-execute "rm -f app/tmp/logs/debug.log"
-execute "rm -f app/tmp/cache/persistent/myapp_*"
-
-execute "cd $PKGDIR"
-execute "zip -r $ZIPFILE.zip $PROJECTNAME"
 
 
 #-- end --
